requirements.txt
numpy
matplotlib
scipy
Pillow
ezdxf
tqdm
